Fix a merge conflict issue with menu
